+++ rental.rb
@@ -1,6 +1,7 @@
 class Rental
   attr_accessor :date
   attr_reader :book, :person
+  @@rental =[]
 
   def initialize(date, book, person)
     @date = date
